Refactor LocationId
Change from Int64 to Int32

+++ src/Carbon.Platform.Abstractions/Hosting/IDisruption.cs
@@ -6,7 +6,7 @@ namespace Carbon.Platform.Disruptions
     {
         long Id { get; }
 
-        long LocationId { get; }
+        int LocationId { get; }
         
         // scope = buckets, ...
