+++ src/Carbon.Platform/Computing/VolumeInfo.cs
@@ -45,7 +45,7 @@ namespace Carbon.Platform.Computing
 
         [IgnoreDataMember]
         [Member("locationId")]
-        public long LocationId { get; }
+        public int LocationId { get; }
 
         ResourceType IResource.ResourceType => ResourceType.Volume;
