+++ src/Carbon.Platform/Data/DatabaseCluster.cs
@@ -43,7 +43,7 @@ namespace Carbon.Platform.Data
 
         [IgnoreDataMember]
         [Member("locationId")]
-        public long LocationId { get; }
+        public int LocationId { get; }
 
         ResourceType IResource.ResourceType => ResourceType.DatabaseCluster;
