+++ src/Carbon.Platform/Data/EncryptionKey.cs
@@ -50,7 +50,7 @@ namespace Carbon.Platform.Data
 
         [IgnoreDataMember]
         [Member("locationId")]
-        public long LocationId { get; }
+        public int LocationId { get; }
 
         ResourceType IResource.ResourceType => ResourceType.EncryptionKey;
