+++ src/Carbon.Platform/Networking/NetworkInterfaceInfo.cs
@@ -63,7 +63,7 @@ namespace Carbon.Platform.Networking
 
         [IgnoreDataMember]
         [Member("locationId")]
-        public long LocationId { get; }
+        public int LocationId { get; }
 
         ResourceType IResource.ResourceType => ResourceType.NetworkInterface;
