+++ src/Carbon.Platform/Networking/NetworkSecurityGroup.cs
@@ -42,7 +42,7 @@ namespace Carbon.Platform.Networking
         [Ascii, StringLength(100)]
         public string ResourceId { get; set; }
         
-        long IManagedResource.LocationId => 0;
+        int IManagedResource.LocationId => 0;
 
         ResourceType IResource.ResourceType => ResourceType.NetworkSecurityGroup;
